Refactor label data type from int to int32 for consistency across datasets

+++ datasets/KITTI.py
@@ -96,7 +96,7 @@ class KITTIDataset(data.Dataset):
         frag2 = tgt_keypts[corr[:, 1]]
         frag1_warp = transform(frag1, gt_trans)
         distance = np.sqrt(np.sum(np.power(frag1_warp - frag2, 2), axis=1))
-        labels = (distance < self.inlier_threshold).astype(np.int)
+        labels = (distance < self.inlier_threshold).astype(np.int32)
         soft_labels = distance * labels.astype(np.float32) + (1. - labels) * 10.
 
         # add random outlier to input data
